Add files via upload
In this project we have used PowerBI to analyse the Call Centre Data. The data cleaning process involved the finding the null and duplicates and removing them in case there are any. New measures are created to show different KPIs as per the requirement. Created new date table from given data and data modelling is done. And the visualizations are done using Bar charts,Map chart,pie chart,tree map for analysing data.